perf: fold approval display helpers into the card to respect the ui-ready module ratchet
