+++ tldw_chatbook/Chat/console_chat_controller.py
@@ -36,10 +36,7 @@ import weakref
 from loguru import logger
 from rich.markup import escape as escape_markup
 
-from tldw_chatbook.Chat.approval_display import (
-    TOOL_DESCRIPTION_CAPTURE_CAP,
-    summarize_arguments,
-)
+from tldw_chatbook.Widgets.Chat_Widgets.chat_approval_card import TOOL_DESCRIPTION_CAPTURE_CAP
 from tldw_chatbook.Character_Chat.emote_directives import (
     CharacterEmoteAssetReference,
     CharacterEmoteRunSnapshot,
