+++ tldw_chatbook/Chat/permission_summary_service.py
@@ -11,7 +11,7 @@ from dataclasses import dataclass, replace
 from typing import Any, Callable, Optional
 
 from tldw_chatbook.Chat.Chat_Functions import chat_api_call, chat_reply_text
-from tldw_chatbook.Chat.approval_display import (
+from tldw_chatbook.Widgets.Chat_Widgets.chat_approval_card import (
     TOOL_DESCRIPTION_CAPTURE_CAP,
     format_context_line,
     summarize_arguments,
